backend/requirements.txt
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.22.0
azure-iot-device==2.14.0
azure-digitaltwins-core==1.2.0
azure-eventhub==1.20.0
azure-cosmos==1.20.0
azure-functions==1.20.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
pandas==2.2.2
numpy==1.24
fastapi==0.113.0
scikit-learn==1.5.1
prophet==1.1.5